MCP/requirements.txt
# --- Core MCP + FastAPI integration ---
fastmcp==0.3.1
fastapi==0.115.0
uvicorn==0.30.6

# --- Database & ORM ---
sqlalchemy==2.0.36
alembic==1.13.3

# --- Password hashing ---
argon2-cffi==23.1.0

# --- Image processing / OCR ---
opencv-python==4.10.0.84
pytesseract==0.3.13
Pillow==10.3.0

# --- Hugging Face / AI utilities ---
transformers==4.45.2
torch==2.4.1
requests==2.32.3

# --- Environment & config ---
python-dotenv==1.0.1

# --- Misc (if you use async + CLI) ---
aiofiles==24.1.0
typer==0.12.4
